requirements.txt
dockerfile-parse
PyYAML
docker
pytest
pytest-cov
mypy
types-PyYAML
types-dockerfile-parse
caseutil
dotenv
click
click-completion>=0.5.2
sentence-transformers>=2.2.0
jsonschema
anthropic
gnureadline